Got input module working

+++ _build/Input.cpp
@@ -5,49 +5,49 @@ namespace jothly
 {
 	bool Input::GetKeyPressed(KeyCode keyCode)
 	{
-		return IsKeyPressed(keyCode);
+		return IsKeyPressed((int) keyCode);
 	}
 
 
 	bool Input::GetKeyReleased(KeyCode keyCode)
 	{
-		return IsKeyReleased(keyCode);
+		return IsKeyReleased((int) keyCode);
 	}
 
 
 	bool Input::GetKeyDown(KeyCode keyCode)
 	{
-		return IsKeyDown(keyCode);
+		return IsKeyDown((int) keyCode);
 	}
 
 
 	bool Input::GetKeyUp(KeyCode keyCode)
 	{
-		return IsKeyUp(keyCode);
+		return IsKeyUp((int) keyCode);
 	}
 
 
 	bool Input::GetMousePressed(MouseCode mouseCode)
 	{
-		return IsMouseButtonPressed(mouseCode);
+		return IsMouseButtonPressed((int) mouseCode);
 	}
 
 
 	bool Input::GetMouseReleased(MouseCode mouseCode)
 	{
-		return IsMouseButtonReleased(mouseCode);
+		return IsMouseButtonReleased((int) mouseCode);
 	}
 
 
 	bool Input::GetMouseButtonDown(MouseCode mouseCode)
 	{
-		return IsMouseButtonDown(mouseCode);
+		return IsMouseButtonDown((int) mouseCode);
 	}
 
 
 	bool Input::GetMouseButtonUp(MouseCode mouseCode)
 	{
-		return IsMouseButtonUp(mouseCode);
+		return IsMouseButtonUp((int) mouseCode);
 	}
